pyproject.toml
[tool.poetry]
name = "repl-nix-stocktrackerpro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
yfinance = "^0.2.43"
pandas = "^2.2.3"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
